+++ src/store/create-store.js
@@ -22,6 +22,7 @@ import mediaSessionMiddleware from '../playback/media-session-middleware';
 import silentHtml5AudioMiddleware from '../playback/silent-html5-audio-middleware';
 import confirmExitMiddleware from '../settings/confirm-exit-middleware';
 import sentryMiddleware from '../sentry/sentry-middleware';
+import resumeAudioContextMiddleware from '../playback/resume-audio-context-middleware';
 
 const createStore = (preloadedState) =>
   _createStore(
@@ -37,6 +38,7 @@ const createStore = (preloadedState) =>
     preloadedState,
     applyMiddleware(
       sentryMiddleware,
+      resumeAudioContextMiddleware,
       anonymousImportMiddleware,
       playbackMiddleware,
       storeUserStateMiddleware,
